% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLike.R
\name{checkLike}
\alias{checkLike}
\title{checkLike}
\usage{
checkLike(str, codes, dict)
}
\arguments{
\item{str}{A character string. Used to match the pattern of 
the first few digits of an ICD10 code.}

\item{codes}{A vector of ICD10 codes as strings. Exclusions}

\item{dict}{A vector of ICD10 codes as strings.}
}
\value{
Returns \code{TRUE} if all strings in \code{dict} starting with 
\code{str} are in \code{codes}, \code{FALSE} otherwise.
}
\description{
checkLike
}
