% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToLike.R
\name{createLIKE}
\alias{createLIKE}
\alias{convertToLike}
\title{Create lists of ICD10 codes to be used in generating SQL queries}
\usage{
createLIKE(codes, dict, n = 2, minGroup = 2)

convertToLike(codes, dict, minGroup = 2, maxLength = 3)
}
\arguments{
\item{codes}{A vector of ICD10 codes as strings. Exclusions.}

\item{dict}{A vector of ICD10 codes as strings.}

\item{n}{The length of starting strings used for grouping ICD10 codes.}

\item{minGroup}{The minimum number of occurrences for a starting string pattern
to be put in the \code{$LIKE} list.}

\item{maxLength}{The maximum length of starting strings considered in \code{$LIKE}.}
}
\value{
Both functions return a list object.
\item{LIKE}{Sub-lists of starting string patterns.}
\item{IN}{All other ICD10 codes in \code{codes} that don't match 
those patterns. }
}
\description{
\code{createLIKE} and \code{convertToLike} create two lists of ICD10 codes. The first one contains 
a sets of common patterns of the first few digits of the ICD10 codes, which can 
then be used in "LIKE" SQL commands. The second list contains ICD10 codes that 
don't share common patterns, which can be put in "IN" SQL commands.
}
\details{
\code{convertToLike} performs similar task to \code{createLIKE}, except 
that instead of using a fixed length, it considers all starting strings with length 
from 1 to \code{maxLength}. 

Note that the ICD10 codes in the output are all from \code{codes}.
}
\examples{
# Exclusions
bpExclICD9 <- c(1400:2089,2300:2399,8000:8399,8500:8549,8600:8699,
                9050:9099,92611,92612,929,9520:9529,9580:9599,3040:3042,
                3044,3054:3057,34460,7292,4210,4211,4219,0380:0389,01,
                7300:7309,7832,78079,7808,2859)
bpExclICD10 <- getICD10(bpExclICD9)

createLIKE(codes = bpExclICD10, dict = ICD9to10$ICD10)

convertToLike(codes = bpExclICD10, dict = ICD9to10$ICD10, maxLength = 4)
}
