% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_CPT.R
\docType{data}
\name{dict_CPT}
\alias{dict_CPT}
\title{Complete range of CPT codes}
\format{A character vector of 82206 elements.}
\source{
\link{https://en.wikipedia.org/wiki/Current_Procedural_Terminology#cite_note-1}
}
\usage{
data(dict_CPT)
}
\description{
\code{dict_CPT} is a vector containing all CPT codes.
}
\details{
There are three types of CPT code: Category I, Category II, and Category III.
\code{dict_CPT[1:80623] are from Category I, which are all numerical. } 
\code{dict_CPT[80624:82014]} are from Category II and all of them are numerical codes 
with an "F" attached at the end. \code{dict_CPT[82015:82206]} are from Category III. 
They are similar to Category II codes except the ending character is "T". Regardless 
of category, all CPT codes have five characters.
}
