% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.I10.df.R
\name{get.I10.df}
\alias{get.I10.df}
\title{Create a table of descriptions matched with groups of ICD-10 codes}
\usage{
get.I10.df(str, minGroup = 2, maxLength = 4)
}
\arguments{
\item{str}{A character string containing ICD-9 codes and descriptions 
in parentheses. Copied from supplement eTable1.}

\item{minGroup}{The minimum number of occurrences for a starting string pattern
to be put in a group.}

\item{maxLength}{The maximum length of starting strings considered in \code{$LIKE}.}
}
\value{
A data frame with two colunms: 
\item{Description}{Description for the codes, extracted from eTable1. }
\item{ICD10}{Groups of ICD-10 codes, represented by the shared starting substring}
}
\description{
Create a table of descriptions matched with groups of ICD-10 codes
}
