% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesc.R
\name{getDesc}
\alias{getDesc}
\title{Add description to ICD-9 codes}
\usage{
getDesc(str)
}
\arguments{
\item{str}{A character string containing ICD-9 codes and descriptions 
in parentheses. Copied from supplement eTable1.}
}
\value{
A data frame, with two columns:
\item{ICD9}{The ICD-9 codes that are extracted from \code{str}. }
\item{Description}{Description in eTable1 matching the ICD-9 code. }
}
\description{
Given a string of codes and description mixed together, clean up the string to get
a dataframe of codes matched with description
}
\examples{
str <- "7802 9921 (syncope), 345xx 7803x (epilepsy or convulsions), 43xx (cerebrovascular diseases, including stroke/TIA and subarachnoid hemorrhage), 800xx-804xx 850xx-854xx 870xx-873xx 9590x 910xx 920xx-921xx (head or face trauma), 78097 781xx 7820 7845x (altered mental status, nervous and musculoskeletal system symptoms, including gait abnormality, meningismus, disturbed skin sensation, speech deficits), V1254 V10xx (personal history of stroke/TIA )"
View(getDesc(str))
}
