% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSQL.R
\name{writeSQL}
\alias{writeSQL}
\title{Create SQL commands}
\usage{
writeSQL(codeList, colName, file = NULL, write.file = TRUE,
  add.periods = FALSE)
}
\arguments{
\item{colName}{The corresponding column name to put in "REGEXP_LIKE" commands.}

\item{file}{The path to output txt file.}

\item{write.file}{A logical value controlling whether or not to write the output
into \code{file}. Default is \code{TRUE}. When \code{write.file} is \code{FALSE}, 
the function will return the output as a character string instead.}

\item{add.periods}{If TRUE, processes each 'IN' statement with \link{\code{addPeriods}}. Defaults to FALSE.}

\item{codelist}{List output from \code{createLIKE} or \code{convertToLike}.}
}
\value{
If write.file=FALSE (default), returns a partial sql query as a string.
Otherwise, writes that string to a text file as specified.
The returned string/file contains a series of "REGEXP_LIKE" and "IN" commands
connected by "OR".
}
\description{
This function creates a series of "REGEXP_LIKE" and "IN" SQL 
commands to query records with certain ICD codes.
}
\details{
This function calls \code{convert2regex} to create "REGEXP_LIKE" 
commands to query codes in \code{codelist$LIKE}, and creates "IN" commands 
to query codes in \code{codelist$IN}
}
\examples{
ICD9 <- c(1400:2089,2300:2399,8000:8399,8500:8549,8600:8699,
                9050:9099,92611,92612,929,9520:9529,9580:9599,3040:3042,
                3044,3054:3057,34460,7292,4210,4211,4219,0380:0389,01,
                7300:7309,7832,78079,7808,2859)
ICD10 <- getICD10(ICD9)
codeList <- convertToLike(codes = ICD10, dict = ICD9to10$ICD10,2,4)

cat(writeSQL(codeList,'DX_CD',file=NULL,write.file=FALSE,add.periods=FALSE))
}
\seealso{
\code{\link{convert2regex}}, \code{\link{createLIKE}}, 
\code{\link{convertToLike}}
}
